// src/components/product/src/Store/types.d.ts
export interface productInterface {
    [key: string]: string | number | boolean | null | undefined
    article?: string
    price?: number
    image?: string
    color?: string
    colorful_temperature?: number
    light_flow?: number
    power?: number
    angle_of_light?: number
    brightness_adjustment?: boolean
}